refactor: index.ts 쪼개기 - 뷰 렌더링, 이벤트 핸들러 함수 분리 xianeml#4
